all: Don't compress the images on 32bit.
pixz seems to have issues compressing the image files on 32bit systems.
Since we will always build official images on 64bit, only set it to compress and sha1sum the compressed file on 64bit.

+++ chromebook-arm-samsung.sh
@@ -336,8 +336,12 @@ rm -rf ${basedir}/kernel ${basedir}/bootp ${basedir}/root ${basedir}/kali-$archi
 
 echo "Generating sha1sum for kali-$1-chromebook.img"
 sha1sum kali-$1-chromebook.img > ${basedir}/kali-$1-chromebook.img.sha1sum
+# Don't pixz on 32bit, there isn't enough memory to compress the images.
+MACHINE_TYPE=`uname -m`
+if [ ${MACHINE_TYPE} == 'x86_64' ]; then
 echo "Compressing kali-$1-chromebook.img"
 pixz ${basedir}/kali-$1-chromebook.img ${basedir}/kali-$1-chromebook.img.xz
 rm ${basedir}/kali-$1-chromebook.img
 echo "Generating sha1sum for kali-$1-chromebook.img.xz"
 sha1sum kali-$1-chromebook.img.xz > ${basedir}/kali-$1-chromebook.img.xz.sha1sum
+fi
