+++ cubox.sh
@@ -213,8 +213,12 @@ rm -rf ${basedir}/kernel ${basedir}/root ${basedir}/kali-$architecture ${basedir
 
 echo "Generating sha1sum for kali-$1-cubox.img"
 sha1sum kali-$1-cubox.img > ${basedir}/kali-$1-cubox.img.sha1sum
+# Don't pixz on 32bit, there isn't enough memory to compress the images.
+MACHINE_TYPE=`uname -m`
+if [ ${MACHINE_TYPE} == 'x86_64' ]; then
 echo "Compressing kali-$1-cubox.img"
 pixz ${basedir}/kali-$1-cubox.img ${basedir}/kali-$1-cubox.img.xz
 rm ${basedir}/kali-$1-cubox.img
 echo "Generating sha1sum for kali-$1-cubox.img.xz"
 sha1sum kali-$1-cubox.img.xz > ${basedir}/kali-$1-cubox.img.xz.sha1sum
+fi
